# api-test/src/generated/swagger.json
{
	"openapi": "3.0.0",
	"components": {
...
		"schemas": {
			"TestResponse": {
				"properties": {
					"id": {
						"type": "string"
					},
					"name": {
						"type": "string"
					},
					"sceneId": {
						"type": "string"
					},
					"state": {
						"type": "string",
						"enum": [
							"active",
							"inactive"
						]
					},
					"createdAt": {
						"type": "string",
						"format": "date-time"
					}
				},
				"required": [
					"id",
					"name",
					"sceneId",
					"state",
					"createdAt"
				],
				"type": "object",
				"additionalProperties": false
			},
			"CreateTestRequest": {
				"properties": {
					"name": {
						"type": "string"
					},
					"sceneId": {
						"type": "string"
					},
					"state": {
						"type": "string",
						"enum": [
							"active",
							"inactive"
						]
					}
				},
				"required": [
					"name",
					"sceneId"
				],
				"type": "object",
				"additionalProperties": false
			},
			"CreateMultipleTestsResponse": {
				"properties": {
					"success": {
						"type": "boolean"
					},
					"message": {
						"type": "string"
					},
					"createdTests": {
						"items": {
							"$ref": "#/components/schemas/TestResponse"
						},
						"type": "array"
					},
					"failedTests": {
						"items": {
							"properties": {
								"error": {
									"type": "string"
								},
								"name": {
									"type": "string"
								}
							},
							"required": [
								"error",
								"name"
							],
							"type": "object"
						},
						"type": "array"
					}
				},
				"required": [
					"success",
					"message",
					"createdTests"
				],
				"type": "object",
				"additionalProperties": false
			},
			"CreateMultipleTestsRequest": {
				"properties": {
					"sceneId": {
						"type": "string"
					},
					"tests": {
						"items": {
							"properties": {
								"state": {
									"type": "string",
									"enum": [
										"active",
										"inactive"
									]
								},
								"name": {
									"type": "string"
								}
							},
							"required": [
								"name"
							],
							"type": "object"
						},
						"type": "array"
					}
				},
				"required": [
					"sceneId",
					"tests"
				],
				"type": "object",
				"additionalProperties": false
			},
			"StartStopTestResponse": {
				"properties": {
					"id": {
						"type": "string"
					},
					"name": {
						"type": "string"
					},
					"sceneId": {
						"type": "string"
					},
					"state": {
						"type": "string",
						"enum": [
							"active",
							"inactive"
						]
					},
					"message": {
						"type": "string"
					}
				},
				"required": [
					"id",
					"name",
					"sceneId",
					"state",
					"message"
				],
				"type": "object",
				"additionalProperties": false
			},
...
		},
		"securitySchemes": {}
	},
	"info": {
		"title": "api-test",
		"version": "1.0.0",
		"license": {
			"name": "ISC"
		},
		"contact": {}
	},
	"paths": {
		"/tests": {
			"get": {
				"operationId": "GetTestsByScene",
				"responses": {
					"200": {
						"description": "Retrieved tests successfully",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/TestResponse"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"description": "Get all tests for a specific scene",
				"tags": [
					"Tests"
				],
				"security": [],
				"parameters": [
					{
						"description": "The scene ID to get tests for",
						"in": "query",
						"name": "sceneId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			},
			"post": {
				"operationId": "CreateTest",
				"responses": {
					"201": {
						"description": "Test created successfully",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/TestResponse"
								}
							}
						}
					}
				},
				"description": "Create a new test",
				"tags": [
					"Tests"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/CreateTestRequest"
							}
						}
					}
				}
			}
		},
		"/tests/bulk": {
			"post": {
				"operationId": "CreateMultipleTests",
				"responses": {
					"201": {
						"description": "Tests created successfully",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/CreateMultipleTestsResponse"
								}
							}
						}
					}
				},
				"description": "Create multiple tests at once",
				"tags": [
					"Tests"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/CreateMultipleTestsRequest"
							}
						}
					}
				}
			}
		},
		"/tests/{testId}/start-stop": {
			"put": {
				"operationId": "StartStopTest",
				"responses": {
					"200": {
						"description": "Test state toggled successfully",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/StartStopTestResponse"
								}
							}
						}
					}
				},
				"description": "Start or stop a test by toggling its state",
				"tags": [
					"Tests"
				],
				"security": [],
				"parameters": [
					{
						"description": "The test ID to start or stop",
						"in": "path",
						"name": "testId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/tests/{testId}": {
			"get": {
				"operationId": "GetTestById",
				"responses": {
					"200": {
						"description": "Retrieved test successfully",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/TestResponse"
								}
							}
						}
					}
				},
				"description": "Get a specific test by ID",
				"tags": [
					"Tests"
				],
				"security": [],
				"parameters": [
					{
						"description": "The test ID to retrieve",
						"in": "path",
						"name": "testId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
...
	},
	"servers": [
		{
			"url": "/"
		}
	]
}